Litecoin: send dialog now more awesome
Now, the send dialog is rather advanced.  In the case of insufficient
fees, it will ask the user if they want to continue, and properly
recreate the sendrequest.  If after the addition of extra fees the
wallet doesn't have enough money, the user is informed of the amount
of the shortfall and asked to manually reconfigure their tx or get
more money.