Add 'bit' precision display
WIth the rising popularity of bits as a denomination, I propose adding bit with two digits after the decimal point. This is very useful, because until the former "µBTC, no digits", this does not result in a loss of information - the user sees his balance down to the last satoshi.