Extracting checkpoint code and setting QR privkeys to time 0
The checkpoints work the same, but I went from using
a helper static method to just instantiating the checkpoint
locally instead.  This gives a little more control if someone
is trying to debug or add features to the checkpointing.  I also
set all QR imported private keys to timestamp 0 to force a FULL
blockchain update when the user selects that option.  Before,
old transactions (before the last checkpoint date) wouldn't be
found ever.

# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
# On branch exp
# Your branch is ahead of 'origin/exp' by 2 commits.
#   (use "git push" to publish your local commits)
#
# Changes to be committed:
#   (use "git reset HEAD <file>..." to unstage)
#
#	modified:   .gitignore
#
# Untracked files:
#   (use "git add <file>..." to include in what will be committed)
#
#	base.iml
#	integration-android/integration-android.iml
#	out/
#	sample-integration-android/sample-integration-android.iml
#	wallet/assets/checkpointslitecoin_month
#	wallet/gen-external-apklibs/
#	wallet/wallet.iml